+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Fixed alt+mouse wheel to only adjust the drawing grid opacity without scrolling vertically.
 - Fixed scrolling off the bottom of the canvas when zoomed out in a tall picture.
+- Fix zoom being broken after resizing the canvas when zoomed in.
 
 ## v3.6.0 (2025-11-28)